requirements.txt
numpy~=2.2.6
python-dotenv~=1.1.1
openai~=1.97.1
tqdm~=4.67.1